pyproject.toml
[project]
name = "n8n_qiongbi_ka"
version = "0.1.0"
description = "HTTP REST API for HTML content parsing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",
    "alembic>=1.13.0",
    "jinja2>=3.1.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "aiohttp>=3.8.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "apscheduler>=3.10.0",
]

